+++ app/Filament/Widgets/RecentPosTransactionsTable.php
@@ -46,9 +46,9 @@ class RecentPosTransactionsTable extends AdvancedTableWidget
                 Tables\Columns\TextColumn::make('grand_total')
                     ->label('Total')
                     ->formatStateUsing(function ($state, Penjualan $record) {
-                        $produkTotal = $record->items->sum(fn($item) => (float) ($item->harga_jual ?? 0) * (int) ($item->qty ?? 0));
-                        $jasaTotal = $record->jasaItems->sum(fn($service) => (float) ($service->harga ?? 0));
-                        $diskon = (float) ($record->diskon_total ?? 0);
+                        $produkTotal = $record->items->sum(fn($item) => (int) ($item->harga_jual ?? 0) * (int) ($item->qty ?? 0));
+                        $jasaTotal = $record->jasaItems->sum(fn($service) => (int) ($service->harga ?? 0));
+                        $diskon = (int) ($record->diskon_total ?? 0);
                         $computed = max(0, ($produkTotal + $jasaTotal) - $diskon);
                         $total = ($state ?? 0) > 0 ? $state : $computed;
